feat: GetGpuConfig API
